test(e2e): one home for the shared ISO-date arithmetic
The auth helper carried its own isoDateDaysAgo because the obvious home for it,
stats-helpers, already imports auth-helpers: importing back would close a cycle
between the two largest helper modules. The arithmetic moves to a leaf module
that imports nothing, stats-helpers re-exports it for the specs that already
name it, and the auth helper computes its onboarding anchor from the same two
functions every other seeding path uses.
